........
r61707 | qwell | 2007-04-20 16:35:27 -0500 (Fri, 20 Apr 2007) | 8 lines

Avoid invalid seqno cycling detection.

Per comment from Dave Troy:
 This adds back in some simple typecasting I had in an earlier version
 which I realize now may be breaking things.

Issue #9554.

........


git-svn-id: https://origsvn.digium.com/svn/asterisk/trunk@61708 65c4cc65-6c06-0410-ace0-fbb531ad65f3

+++ main/rtp.c
@@ -1287,7 +1287,7 @@ struct ast_frame *ast_rtp_read(struct ast_rtp *rtp)
 		/* Schedule transmission of Receiver Report */
 		rtp->rtcp->schedid = ast_sched_add(rtp->sched, ast_rtcp_calc_interval(rtp), ast_rtcp_write, rtp);
 	}
-	if ( rtp->lastrxseqno - seqno  > 100) /* if so it would indicate that the sender cycled; allow for misordering */
+	if ( (int)rtp->lastrxseqno - (int)seqno  > 100) /* if so it would indicate that the sender cycled; allow for misordering */
 		rtp->cycles += RTP_SEQ_MOD;
 
 	rtp->lastrxseqno = seqno;
